% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.multibergm.R
\name{GetNetStats}
\alias{GetNetStats}
\title{Compute network statistics}
\usage{
GetNetStats(object, formula, whichStats)
}
\arguments{
\item{object}{A network or list of networks}

\item{formula}{The ERGM formula containing the summary statistics to be
computed}

\item{whichStats}{A string specifying which network statistics to be
computed ("all", "model", "gof", or "other").}
}
\description{
GetNetStats is an internal function used to compute network statistics.
}
