% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.multibergm.R
\name{gof.multibergm}
\alias{gof.multibergm}
\title{Goodness-of-fit diagnostics for multibergm}
\usage{
\method{gof}{multibergm}(
  object,
  coefs = NULL,
  sample_size = 100,
  aux_iters = 1.5 * object$control$aux_iters,
  burn_in = 0L,
  thin = 1L
)
}
\arguments{
\item{object}{A \link{multibergm} object}

\item{coefs}{Optional set of model coefficients to use for network
simulation. The default is to use the posterior samples of the
population-level parameter.}

\item{thin}{Amount of thinning to apply to the posterior samples}

\item{sampleSize}{Number of networks to be simulated and compared to the
observed networks.}

\item{auxIters}{Number of iterations used for network simulation.}
}
\value{
Outputs bar plots (for observed networks) overlayed with ribbons
  (for simulated network) of degree, minimum geodesic distances,
  and edge-wise shared partner distributions.
}
\description{
Function to calculate summaries for degree, minimum geodesic distances,
and edge-wise shared partner distributions to diagnose the Bayesian
goodness-of-fit of exponential random graph models fit to multiple
networks.
}
