% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_multibergm.R
\name{initialise_multibergm}
\alias{initialise_multibergm}
\title{Initialise a multibergm fit}
\usage{
initialise_multibergm(
  formula,
  constraints = ~.,
  groups = NULL,
  prior = set_priors(formula, groups),
  init = set_init(formula, prior, groups),
  control = control_multibergm(formula, constraints)
)
}
\description{
Initialise a multibergm fit by specifying the ERGM formula, group
memberships, prior hyperparameters, initial values and tuning parameters for 
the MCMC.
}
