% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krackfr-data.R
\docType{data}
\name{krackfr}
\alias{krackfr}
\title{Friendship networks from Krackhardt.}
\format{
A list with a set of 21 networks.
}
\usage{
krackfr
}
\description{
A dataset containing the friendship networks.
}
\keyword{datasets}
