% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multibergm_updates.R
\name{multi_var_update}
\alias{multi_var_update}
\title{MCMC update for multiple-group multi-BERGM}
\usage{
multi_var_update(curr, prior, groups, proposals, control)
}
\arguments{
\item{prior}{A list of explicit prior specifications.}

\item{groups}{A vector of group memberships}

\item{control}{A list of parameters set by \code{\link{control_multibergm}}}
}
\value{
A list of the updated values of the model parameters and the
  acceptance counts for the exchange updates.
}
\description{
Function used to perform one exchange-within-Gibbs MCMC update for
all the parameters in a multiple-group multi-BERGM. This applies a standard
Gibbs update for the population-level parameters and the network-level
covariance parameter before using the exchange algorithm within an
Ancillarity-Sufficiency Interweaving Strategy (ASIS) to update the
remaining parameters.
}
