% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nets}
\alias{nets}
\title{Ten synthetic networks}
\format{
A list of networks with 40 nodes.
}
\usage{
nets
}
\description{
A dataset of ten networks simulated from the same exponential random graph
model with different model parameters.
}
\keyword{datasets}
