% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.multibergm.R
\name{print.multibergm}
\alias{print.multibergm}
\title{S3 print method for a multibergm object}
\usage{
\method{print}{multibergm}(x, ...)
}
\arguments{
\item{x}{A \code{\link{multibergm}} object}

\item{...}{Further arguments passed to other methods}
}
\description{
Automatically called when an object of class \code{\link{multibergm}} is
printed. Currently, summarizes the number of networks, the number of groups,
the model formula and the model constraints.
}
