% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_priors.R
\name{set_priors}
\alias{set_priors}
\title{Set priors for multibergm}
\usage{
set_priors(
  ergm_formula,
  model_matrix,
  control,
  hyper_model_matrix = NULL,
  prior = list()
)
}
\arguments{
\item{ergm_formula}{An \R \code{\link{formula}} object, of the form
\code{y ~ <model terms>}, where \code{y} is a
\code{\link[network]{network}} object or a
\code{\link[ergm]{network.list}} object.}

\item{control}{A list of parameters set by \code{\link{control_multibergm}}}

\item{prior}{A list of explicit prior specifications.}
}
\description{
This is a utility function to set priors for a multibergm. When not
explicitly specified, set_priors() uses default (vague) priors. A
validity check is also performed to ensure the compatibility of the priors
with the model.
}
