% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_proposals.R
\name{update_proposals}
\alias{update_proposals}
\title{Update Metropolis random-walk covariance parameters}
\usage{
update_proposals(proposals, accept_rate, params, control)
}
\arguments{
\item{proposals}{Current set of proposals}

\item{accept_rate}{Acceptance rates since the last proposal adaptation}

\item{params}{Posterior samples}

\item{control}{A list of parameters set by \code{\link{control_multibergm}}}
}
\description{
This is an internal function used to update the covariance parameters for the
mu_pop and theta proposals.
}
