% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_updates.R
\name{mcmc_updates}
\alias{mcmc_updates}
\alias{exchange_update}
\alias{mean_update}
\alias{cov_update}
\alias{ergm_wrapper}
\title{Internal functions for MCMC updates}
\usage{
exchange_update(curr, prop, delta, priorCov, priorMean = NULL,
  netLabels = NULL)

mean_update(obsData, obsCov, priorMean, priorCov)

cov_update(obsData, priorDf, priorScale, obsMean, dataLabels = NULL,
  currCov = NULL)

ergm_wrapper(coefs, control)
}
\arguments{
\item{curr}{A vector or matrix of current mean parameter values.}

\item{prop}{A vector or matrix of proposed mean parameter values}

\item{delta}{Change in summary statistics as produced by
\code{\link{ergm_wrapper}}}

\item{priorCov}{Prior covariance of parameter}

\item{priorMean}{Prior mean of parameter}

\item{netLabels}{Labels specifying which networks to associate with
each row of the the parameter matrices}

\item{obsData}{Matrix of observed data}

\item{obsCov}{Fixed (known) covariance of observed data}

\item{priorDf}{Prior degrees of freedom in Inverse-Wishart}

\item{priorScale}{Prior scale matrix in Inverse-Wishart}

\item{obsMean}{Fixed (known) mean of observed data}

\item{dataLabels}{Labels to associate each observation with a grouping}

\item{currCov}{Current value of covariance parameter}

\item{coefs}{Matrix of coefficients to simulate ERGs from}

\item{control}{A list of parameters set by \code{\link{control_multibergm}}
containing settings for the ERGM simulations}
}
\description{
Functions used internally by to perform Gibbs or exchange updates.
}
\section{Functions}{
\itemize{
\item \code{mean_update}: Gibbs update of a mean parameter

\item \code{cov_update}: Gibbs update of a covariance parameter

\item \code{ergm_wrapper}: Gibbs update of a mean parameter
}}

