% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multibergm.R
\name{plot.multibergm}
\alias{plot.multibergm}
\title{Plot BERGM posterior output}
\usage{
\method{plot}{multibergm}(object, param = "mu", burn_in = 0L, thin = 1L, ind = 1L, ...)
}
\arguments{
\item{object}{A \link{multibergm} object}

\item{param}{Multibergm parameter to be summarised}

\item{burn_in}{Amount of burn-in to remove from the start of posterior
samples (pre-thinning)}

\item{thin}{Amount of thinning to apply to the posterior samples}

\item{ind}{Indices to subset when param is theta or mu_pop}

\item{...}{Additional parameters to be passed on to lower-level functions.}

\item{x}{A \link{multibergm} object}
}
\value{
Outputs a density plot, trace plot, and autocorrelation plot for
  each of the model parameters.
}
\description{
This function creates MCMC diagnostic plots for \code{multibergm} objects.
}
