% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multibergm.R
\name{plot.multibergm}
\alias{plot.multibergm}
\title{Plot BERGM posterior output}
\usage{
\method{plot}{multibergm}(object, param = "mu", burn_in = 0L, thin = 1L)
}
\arguments{
\item{object}{A \link{multibergm} object}

\item{param}{Multibergm parameter to be summarised}

\item{thin}{Amount of thinning to apply to the posterior samples}
}
\value{
Outputs a density plot, trace plot, and autocorrelation plot for
  each of the model parameters.
}
\description{
This function creates MCMC diagnostic plots for \code{multibergm} objects.
}
