% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.multibergm.R
\name{summary.multibergm}
\alias{summary.multibergm}
\title{Summarising multibergm model fits}
\usage{
\method{summary}{multibergm}(object, param = "mu", thin = 1L, burn_in = 0L, ...)
}
\arguments{
\item{object}{A \link{multibergm} object}

\item{param}{Multibergm parameter to be summarised}

\item{thin}{Amount of thinning to apply to the posterior samples}

\item{burn_in}{Amount of burn-in to remove from the start of posterior
samples (pre-thinning)}

\item{...}{Additional parameters to be passed on to lower-level functions.}
}
\value{
The function computes and prints posterior means and quantiles for
  the specified parameter, as well as the acceptance rates for parameters
  updated via the exchange algorithm.
}
\description{
\link{summary} method for \link{multibergm} fits
}
